Add the ability to send control messages using CamStream.
In this approach, the main camera loop also checks for
user input in another mp.Queue, and if present, writes
everything as a stream of control messages before proceeding
to the next frame. The assumption is that user input rate won't
usually match frame rate, and the delay would be tolerable.